Support for websocketsApiRouteSelectionExpression
In the form of $request.body.x.y.z
