+++ manual_test_websocket/manual_test_websocket_RouteSelection/package.json
@@ -13,7 +13,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "aws-sdk": "^2.469.0"
+    "aws-sdk": "^2.472.0"
   },
   "devDependencies": {
     "aws4": "^1.8.0",
