+++ manual_test_websocket/manual_test_websocket_RouteSelection/scripts/deploy_to_aws.sh
