+++ manual_test_websocket/manual_test_websocket_RouteSelection/scripts/serverless.aws.yml
