+++ manual_test_websocket/manual_test_websocket_RouteSelection/scripts/serverless.offline.yml
@@ -1,0 +1,7 @@
+plugins:
+  - serverless-offline
+
+custom:
+  serverless-offline:
+    port: 3004
+
