--- manual_test_websocket/manual_test_websocket_authorizer/scripts/serverless.offline.yml
@@ -1,4 +1,0 @@
-plugins:
-  - serverless-offline
-
-
